fix: adjust feedback button display
